Rename package from expose to read to be more meaningful in REST
