+++ src/main/java/spyra/lukasz/pokerestapi/PokeRestApiApplication.java
@@ -8,7 +8,7 @@ import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.client.RestOperations;
-import spyra.lukasz.pokerestapi.expose.exceptions.RestTemplateResponseErrorHandler;
+import spyra.lukasz.pokerestapi.read.exceptions.RestTemplateResponseErrorHandler;
 
 @EnableCaching
 @SpringBootApplication
